+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Next
 
 <<<<<<< HEAD
-- Feature (`@grafana/faro-core`): sourcemap uploads - add `bundleId` to the `MetaApp` `Meta` object (#476).
-=======
+
+- # Feature (`@grafana/faro-core`): sourcemap uploads - add `bundleId` to the `MetaApp` `Meta` object (#476).
 - Fix (`@grafana/faro-web-sdk`): Faro updates sessions in an infinite loop if DOM Storage is not
   available (#519).
 
@@ -33,7 +33,8 @@
 
 ## 1.3.8
 
->>>>>>> c33caee495826c6c73443114c791ef2158af15ef
+> > > > > > > c33caee495826c6c73443114c791ef2158af15ef
+
 - Deps (`@grafana/faro-web-tracing`, `@grafana/faro-core`): Update OpenTelemetry dependencies (#475).
 
 ## 1.3.7