+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+- Feat(`@grafana/faro-web-sdk`): Add responseStatus to performance events (#526).
 - Fix (`@grafana/faro-web-sdk`): Faro updates sessions in an infinite loop if DOM Storage is not
   available (#519).
 - Feat (`@grafana/faro-react`): Support instrumenting React Router v6 data routers (#518).