+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Next
 
 - Feat (`@grafana/faro-core`): sourcemap uploads - add `bundleId` to the `MetaApp` `Meta` object (#476).
-- Feat (`@grafana/faro-web-sdk`): sourcemap uploads - add `bundleid` to each stack frames if uploaded source maps are used (#476).
+- Feat (`@grafana/faro-web-sdk`): sourcemap uploads - add `bundleid` to each stack frames (#476).
 - Feat (`@grafana/faro-web-sdk`): Add responseStatus to performance events (#526).
 - Fix (`@grafana/faro-web-sdk`): Faro updates sessions in an infinite loop if DOM Storage is not
   available (#519).